feat!: Focus tracking for field validation and gloval validation marking.
Updating to allow the editor to be marked for global validation. Fields will also show validation if the field has lost focus indicating that the user is done with the input.

The main goal is to provide in context validation after user interaction in a non-intrusive way while also supporting whole form validation (ex: on 'submit').

Breaking change: zoneToKey is reworked to be zones with a structured value.

fixes #44