feat: Delay validation on a form until it is marked for validation. (#56)
On long forms the user experience is enhanced by providing validation after the user has visited a field. But for shorter forms or other circumstances having immediate validation can be jarring for the experience. Using the delayed config allows the validation to be delayed for the entire editor until the editor has been marked for validation.